Move testing structs to separate file
